One Front Door/SourceCode/MerchAngular4/dpsg.merchandising.web/src/app/pl-merch/pl-merch-item/pl-merch-item.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { PlMerchItemComponent } from './pl-merch-item.component';

describe('Component: PlMerchItem', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [PlMerchItemComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([PlMerchItemComponent],
      (component: PlMerchItemComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(PlMerchItemComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(PlMerchItemComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-pl-merch-item></app-pl-merch-item>
  `,
  directives: [PlMerchItemComponent]
})
class PlMerchItemComponentTestController {
}

